Split sandbox.ts into src/sandbox/, one function per file
